Removed naive try/except on unicode builtin to determine Python version
In the case a user has defined a builtin for legacy compatibility, the try/except method naively determines the user must be running Python2 as unicode exists and the try does not raise an Exception.
This makes later internal calls to zone.replace() fail with a TypeError in Python 3 as it attempts to replace str types in a bytes-like object.